feat: implement asynchronous processing
Using threads and a mutex, we can process Open Graph URLs asynchronously.

Now we can do batch processing faster.

However, there is no guarantee that the webpage can immediately give the
processing result of the latest submission.

For the latest results, we need to refresh the page manually after the system
finishes processing.